+++ .storybook/main.ts
@@ -16,5 +16,6 @@ const config: StorybookConfig = {
 	docs: {
 		autodocs: 'tag',
 	},
+	staticDirs: ['../public'],
 }
 export default config
