+++ src/widgets/Button/Button.stories.ts
@@ -5,7 +5,7 @@ import { Button } from './Button'
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 const meta = {
-	title: 'Example/Button',
+	title: 'Kernel/Button',
 	component: Button,
 	parameters: {
 		// Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/configure/story-layout
